Display ships and attacks on gameboard
Ships and attacks, both hits and misses, are now
rendered on gameboard as blue, orange and red
coloured cells respectively.
Also corrected function for placing ships which
was filling each row with the same object
reference resulting in the whole row being
assigned the ship instead of only the desired
cell/s. Updated test to confirm the cells next to
the cell where the ship is being placed are not
inadvertently assigned the ship.